requirements.txt
# LLM Stack Manager - Arquitectura Local Nativa
# Dependencias simplificadas (sin Docker, sin DB)

rich>=13.7.0          # CLI moderna con colores y tablas
requests>=2.31.0      # HTTP requests para Ollama API
pyyaml>=6.0.0         # Configuración YAML
pytest>=7.4.0         # Testing
pytest-asyncio>=0.21.0